Separation of metafunction and runtime regex code.
